export workflow metadata to separate file
